+++ docs/conf.py
@@ -16,10 +16,10 @@ sys.path.insert(0, str(Path(__file__).parent.parent / "src"))
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
 project = "MD2PDF"
-copyright = "2024, MPS Metalmind AB"
+copyright = "2025, MPS Metalmind AB"
 author = "MPS Metalmind AB"
-release = "1.0.0"
-version = "1.0.0"
+release = "1.1.0"
+version = "1.1.0"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration